Add API routes for Git and Anthropic functionalities: implement handlers for file ingestion, retrieval, and Python code execution
