+++ frontend/src/components/ClaudeNode.vue
@@ -12,6 +12,16 @@
         </select>
       </div>
   
+      <!-- Predefined System Prompt Dropdown -->
+      <div class="input-field">
+        <label for="system-prompt-select" class="input-label">Predefined System Prompt:</label>
+        <select id="system-prompt-select" v-model="selectedSystemPrompt" class="input-select">
+          <option v-for="(prompt, key) in systemPromptOptions" :key="key" :value="key">
+            {{ prompt.role }}
+          </option>
+        </select>
+      </div>
+  
       <!-- System Prompt -->
       <div class="input-field">
         <label :for="`${data.id}-system_prompt`" class="input-label">System Prompt (Optional):</label>
@@ -42,7 +52,7 @@
   </template>
   
   <script setup>
-  import { ref, computed, onMounted, nextTick } from 'vue'
+  import { ref, computed, onMounted, nextTick, watch } from 'vue'
   import { Handle, useVueFlow } from '@vue-flow/core'
   import { NodeResizer } from '@vue-flow/node-resizer'
   const { getEdges, findNode } = useVueFlow()
@@ -50,7 +60,40 @@
   const emit = defineEmits(['update:data', 'resize', 'disable-zoom', 'enable-zoom'])
   const showApiKey = ref(false)
   
-  // On mount, attach our run() method to the node’s data if not already set.
+  // New: Predefined System Prompt Options & selection
+  const selectedSystemPrompt = ref("friendly_assistant");
+  const systemPromptOptions = {
+    friendly_assistant: {
+        role: "Friendly Assistant",
+        system_prompt: "You are a helpful, friendly, and knowledgeable general-purpose AI assistant. You can answer questions, provide information, engage in conversation, and assist with a wide variety of tasks. Be concise in your responses when possible, but prioritize clarity and accuracy. If you don't know something, admit it. Maintain a conversational and approachable tone."
+    },
+    search_assistant: {
+        role: "Search Assistant",
+        system_prompt: "You are a helpful assistant that specializes in generating effective search engine queries. Given any text input, your task is to create one or more concise and relevant search queries that would be likely to retrieve information related to that text from a search engine (like Google, Bing, etc.). Consider the key concepts, entities, and the user's likely intent. Prioritize clarity and precision in the queries."
+    },
+    research_analyst: {
+        role: "Research Analyst",
+        system_prompt: "You are a skilled research analyst with deep expertise in synthesizing information. Approach queries by breaking down complex topics, organizing key points hierarchically, evaluating evidence quality, providing multiple perspectives, and using concrete examples. Present information in a structured format with clear sections, use bullet points for clarity, and visually separate different points with markdown. Always cite limitations of your knowledge and explicitly flag speculation."
+    },
+    creative_writer: {
+        role: "Creative Writer",
+        system_prompt: "You are an exceptional creative writer. When responding, use vivid sensory details, emotional resonance, and varied sentence structures. Organize your narratives with clear beginnings, middles, and ends. Employ literary techniques like metaphor and foreshadowing appropriately. When providing examples or stories, ensure they have depth and authenticity. Present creative options when asked, rather than single solutions."
+    },
+    code_expert: {
+        role: "Programming Expert",
+        system_prompt: "You are a senior software developer with expertise across multiple programming languages. Present code solutions with clear comments explaining your approach. Structure responses with: 1) Problem understanding 2) Solution approach 3) Complete, executable code 4) Explanation of how the code works 5) Alternative approaches. Include error handling in examples, use consistent formatting, and provide explicit context for any code snippets. Test your solutions mentally before presenting them."
+    },
+    teacher: {
+        role: "Educational Expert",
+        system_prompt: "You are an experienced teacher skilled at explaining complex concepts. Present information in a structured, progressive manner from foundational to advanced. Use analogies and examples to connect new concepts to familiar ones. Break down complex ideas into smaller components. Incorporate multiple formats (definitions, examples, diagrams described in text) to accommodate different learning styles. Ask thought-provoking questions to deepen understanding. Anticipate common misconceptions and address them proactively."
+    },
+    data_analyst: {
+        role: "Data Analysis Expert",
+        system_prompt: "You are a data analysis expert. When working with data, focus on identifying patterns and outliers, considering statistical significance, and exploring causal relationships vs. correlations. Present your analysis with a clear narrative structure that connects data points to insights. Use hypothetical data visualization descriptions when relevant. Consider alternative interpretations of data and potential confounding variables. Clearly communicate limitations and assumptions in any analysis."
+    }
+  };
+  
+  // On mount, attach our run() method to the node's data if not already set.
   onMounted(() => {
     if (!props.data.run) {
       props.data.run = run
@@ -162,13 +205,13 @@
         hasOutputs: true,
         inputs: {
           api_key: '',
-          model: 'claude-3-5-sonnet-20241022',
+          model: 'claude-3-7-sonnet-latest',
           system_prompt: '',
           user_prompt: 'Hello, Claude!',
           max_tokens: 1024
         },
         outputs: { response: '' },
-        models: ['claude-3-7-sonnet-latest', 'claude-3-5-sonnet-20241022', 'claude-3-5-haiku-latest'],
+        models: ['claude-3-7-sonnet-latest', 'claude-3-5-sonnet-latest', 'claude-3-5-haiku-latest'],
         style: {
           border: '1px solid #666',
           borderRadius: '12px',
@@ -222,6 +265,13 @@
   
   const handleTextareaMouseEnter = () => emit('disable-zoom')
   const handleTextareaMouseLeave = () => emit('enable-zoom')
+  
+  // Update the system prompt textbox when the dropdown selection changes
+  watch(selectedSystemPrompt, (newKey) => {
+    if (systemPromptOptions[newKey]) {
+      system_prompt.value = systemPromptOptions[newKey].system_prompt;
+    }
+  }, { immediate: true });
   </script>
   
   <style scoped>
